Progress bug fixed

+++ BasicRenderer/Raytracer.cpp
@@ -29,7 +29,8 @@ namespace BasicRenderer
 		//Top-left
 		for (uint y = 0u; y < height; y++)
 		{
-			std::cout << "Progress: " << static_cast<int>(std::roundf((y / height) * 100.f)) << "% \r";
+
+			std::cout << "Progress: " << static_cast<uint>(std::roundf((static_cast<float>(y) / height) * 100.f)) << "% \r";
 
 			for (uint x = 0u; x < width; x++)
 			{
